fix: address CLI code review findings for F21.4
- Fix critical bug: send "default_branch" instead of "branch" in
  github_connect (field name mismatch caused --branch to be silently ignored)
- Fix error code mismatch: "GITHUB_REPO_NOT_ACCESSIBLE" to match API
- Extract shared require_auth() and expect_str_field() to commands/mod.rs
- Replace unwrap_or("") on app_id with expect_str_field (fail loudly)
- Replace unwrap_or_default() on releases list with explicit parse error
- Replace unwrap_or("unknown") on promote tag with expect_str_field
- Simplify github_disconnect (remove redundant status check)
- Guard SHA slicing with is_ascii() check to prevent panic
